// starting-point/libs/adf/enterprise/adf-hx-content-services/ui/src/lib/components/actions/content-delete/configs/delete-dialog.interface.ts
/*
 * Copyright © 2005 - 2021 Alfresco Software, Ltd. All rights reserved.
 *
 * License rights for this program may be obtained from Alfresco Software, Ltd.
 * pursuant to a written agreement and any use of this program without such an
 * agreement is prohibited.
 */

import { Document } from '@hylandsoftware/hxcs-js-client';

export interface ContentDeleteDialogData {
    documents: Document[];
    shouldRedirect: boolean;
    refererURL: string;
}

export interface ContentDeletedDialogResponse {
    success: boolean;
    docId: string;
    error?: string;
}